+++ FirebaseAI/Sources/Types/Public/Live/LiveSession.swift
@@ -59,7 +59,6 @@ public final class LiveSession: Sendable {
   ///   - audio: Raw 16-bit PCM audio at 16Hz, used to update the model on the client's
   ///     conversation.
   public func sendAudioRealtime(_ audio: Data) async {
-    // TODO: (b/443984790) address when we add RealtimeInputConfig support
     let message = BidiGenerateContentRealtimeInput(
       audio: InlineData(data: audio, mimeType: "audio/pcm")
     )
@@ -94,6 +93,37 @@ public final class LiveSession: Sendable {
   ///   - text: Text content to append to the current client's conversation.
   public func sendTextRealtime(_ text: String) async {
     let message = BidiGenerateContentRealtimeInput(text: text)
+    await service.send(.realtimeInput(message))
+  }
+
+  /// Manually marks the start of user activity, using the realtime API.
+  ///
+  /// The start of user activity is effectively the start of a user's turn, but depending on the
+  /// configuration defined in ``RealtimeInputConfig``, it may not be interpreted as an
+  /// interruption. An example of the start of user activity could be the user speaking
+  /// (not silence).
+  ///
+  /// Should be followed with a call to ``LiveSession/sendStopActivityRealtime()``; after all the
+  /// data has been sent for the user's turn.
+  ///
+  /// Only required when automatic activity detection is disabled via
+  /// ``ActivityDetectionConfig/disabled()``.
+  public func sendStartActivityRealtime() async {
+    let message = BidiGenerateContentRealtimeInput(activityStart: .init())
+    await service.send(.realtimeInput(message))
+  }
+
+  /// Manually marks the end of user activity, using the realtime API.
+  ///
+  /// The end of user activity is effectively the end of a user's turn, and signals that the model
+  /// can start sending responses.
+  ///
+  /// Should follow after a previous call to ``LiveSession/sendStartActivityRealtime()``.
+  ///
+  /// Only required when automatic activity detection is disabled via
+  /// ``ActivityDetectionConfig/disabled()``.
+  public func sendStopActivityRealtime() async {
+    let message = BidiGenerateContentRealtimeInput(activityEnd: .init())
     await service.send(.realtimeInput(message))
   }